requirements.txt
django
django-prose
mysqlclient
requests